+++ src/main/kotlin/dev/jamiecraane/domain/mathfunctions/MathematicalFunctions.kt
@@ -5,7 +5,7 @@ import kotlin.math.pow
 
 //todo add combo picker to select a function. Add all of the following functions: https://www.mathsisfun.com/sets/functions-common.html
 interface MathematicalFunction {
-    fun evaluate(x: Float): Float
+    fun evaluate(x: Double): Double
 
     val label: String
